refactor: environment
