+++ backend/src/acceptanceTest/java/wooteco/prolog/AcceptanceSteps.java
@@ -1,9 +1,7 @@
 package wooteco.prolog;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.Profile;
 
-@Profile("acceptance")
 public class AcceptanceSteps {
 
     @Autowired
