+++ backend/src/acceptanceTest/java/wooteco/prolog/fixtures/GithubTestController.java
@@ -1,13 +1,11 @@
 package wooteco.prolog.fixtures;
 
-import org.springframework.context.annotation.Profile;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import wooteco.prolog.login.application.dto.GithubAccessTokenRequest;
 import wooteco.prolog.login.application.dto.GithubAccessTokenResponse;
 import wooteco.prolog.login.application.dto.GithubProfileResponse;
 
-@Profile("acceptance")
 @RestController
 public class GithubTestController {
