+++ backend/src/main/resources/application-local.yml
@@ -1,3 +1,6 @@
+server:
+  port: 5000
+
 spring:
   jpa:
     generate-ddl: true
@@ -7,6 +10,26 @@ spring:
       ddl-auto: create-drop
     properties:
       hibernate:
-        show_sql: false
         format_sql: true
+        show_sql: false
         use_sql_comment: false
+
+security:
+  jwt:
+    token:
+      secret-key: secret_key
+      expire-length: 3600000
+
+github:
+  client:
+    id: f91b56445e08d44adb76
+    secret: f02769f780ed3d40f9db9283f5b7cc79ecf9074e
+  url:
+    access-token: https://github.com/login/oauth/access_token
+    profile: https://api.github.com/user
+
+logging:
+  level:
+    org:
+      hibernate:
+        type: trace
