+++ backend/src/test/java/wooteco/prolog/studylog/posttag/application/PostTagServiceTest.java
@@ -4,10 +4,6 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.annotation.DirtiesContext;
-import org.springframework.test.context.TestConstructor;
-import org.springframework.test.context.TestConstructor.AutowireMode;
 import org.springframework.transaction.annotation.Transactional;
 import wooteco.prolog.login.application.dto.GithubProfileResponse;
 import wooteco.prolog.member.application.MemberService;
@@ -19,17 +15,15 @@ import wooteco.prolog.studylog.application.PostTagService;
 import wooteco.prolog.studylog.application.dto.*;
 import wooteco.prolog.studylog.domain.PostTag;
 import wooteco.prolog.studylog.domain.Tag;
+import wooteco.support.utils.IntegrationTest;
 
 import java.util.Arrays;
 import java.util.List;
-import java.util.stream.Stream;
 
 import static java.util.stream.Collectors.toList;
 import static org.assertj.core.api.Assertions.assertThat;
 
-@SpringBootTest
-@DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
-@TestConstructor(autowireMode = AutowireMode.ALL)
+@IntegrationTest
 class PostTagServiceTest {
 
     @Autowired
@@ -51,7 +45,7 @@ class PostTagServiceTest {
         MissionResponse mission = missionService.create(new MissionRequest("미션 이름", level.getId()));
 
         this.member = memberService
-            .findOrCreateMember(new GithubProfileResponse("이름", "별명", "1", "image"));
+                .findOrCreateMember(new GithubProfileResponse("이름", "별명", "1", "image"));
     }
 
     @DisplayName("포스트 태그가 등록되면 포스트 태그를 찾아올 수 있는지 확인한다.")
@@ -70,8 +64,8 @@ class PostTagServiceTest {
         List<String> tagNames = Arrays.asList("태그1", "태그2", "태그3", "태그5", "태그6");
 
         List<String> expectedTagNames = tags.stream()
-            .map(TagResponse::getName)
-            .collect(toList());
+                .map(TagResponse::getName)
+                .collect(toList());
 
         assertThat(expectedTagNames).containsExactlyElementsOf(tagNames);
     }
@@ -79,16 +73,16 @@ class PostTagServiceTest {
     @SafeVarargs
     private final List<PostResponse> addTagRequestToPost(List<TagRequest>... tagRequests) {
         List<PostRequest> posts = Arrays.stream(tagRequests)
-            .map(it -> new PostRequest("이름", "별명", 1L, it))
-            .collect(toList());
+                .map(it -> new PostRequest("이름", "별명", 1L, it))
+                .collect(toList());
 
         return postService.insertPosts(member, posts);
     }
 
     private List<TagRequest> createTagRequests(String... tags) {
         return Arrays.stream(tags)
-            .map(TagRequest::new)
-            .collect(toList());
+                .map(TagRequest::new)
+                .collect(toList());
     }
 
     @DisplayName("태그를 기반으로 포스트 태그를 조회할 수 있는지 확인")
@@ -100,13 +94,13 @@ class PostTagServiceTest {
         List<TagRequest> tagRequests2 = createTagRequests("태그1", "태그2");
         List<TagRequest> tagRequests3 = createTagRequests("태그1", "태그2", "태그3");
         List<PostResponse> postResponses =
-            addTagRequestToPost(tagRequests1, tagRequests2, tagRequests3);
+                addTagRequestToPost(tagRequests1, tagRequests2, tagRequests3);
 
         //when
         List<Tag> insertedTags = postResponses.stream()
-            .flatMap(postResponse -> postResponse.getTags().stream())
-            .map(tagResponse -> new Tag(tagResponse.getId(), tagResponse.getName()))
-            .collect(toList());
+                .flatMap(postResponse -> postResponse.getTags().stream())
+                .map(tagResponse -> new Tag(tagResponse.getId(), tagResponse.getName()))
+                .collect(toList());
 
         List<PostTag> postTags = postTagService.findByTags(insertedTags);
 
@@ -117,9 +111,9 @@ class PostTagServiceTest {
 
     private long countTag(List<PostTag> postTags, String tag) {
         return postTags.stream()
-            .map(PostTag::getTag)
-            .map(Tag::getName)
-            .filter(name -> name.equals(tag))
-            .count();
+                .map(PostTag::getTag)
+                .map(Tag::getName)
+                .filter(name -> name.equals(tag))
+                .count();
     }
 }
