+++ backend/src/test/resources/application.yml
@@ -1,31 +1,3 @@
 spring:
-  jpa:
-    generate-ddl: true
-    database: h2
-    database-platform: org.hibernate.dialect.H2Dialect
-    hibernate:
-      ddl-auto: create-drop
-    properties:
-      hibernate:
-        show_sql: true
-        format_sql: true
-        use_sql_comment: true
-logging:
-  level:
-    org:
-      hibernate:
-        type: trace
-
-github:
-  client:
-    id: f91b56445e08d44adb76
-    secret: f02769f780ed3d40f9db9283f5b7cc79ecf9074e
-  url:
-    access-token: http://localhost:4000/github/login/oauth/access_token
-    profile: http://localhost:4000/github/user
-
-security:
-  jwt:
-    token:
-      secret-key: xpzmzhtmvmfhrmvmfhwprxm
-      expire-length: 3600000
+  profiles:
+    active: test
